Consolidate unit tests into single file
